+++ tests/Grid/GridConfigTest.php
@@ -3,6 +3,7 @@
 namespace Kitpages\DataGridBundle\Grid;
 
 use Doctrine\ORM\QueryBuilder;
+use Kitpages\DataGridBundle\Paginator\PaginatorConfig;
 use PHPUnit\Framework\TestCase;
 
 class GridConfigTest extends TestCase
@@ -70,5 +71,15 @@ class GridConfigTest extends TestCase
         $this->assertCount(0, $this->gridConfig->getFieldListByTag('gloubi'));
         $field = $this->gridConfig->getFieldListByTag('biz')[0];
         $this->assertEquals('f3', $field->getFieldName());
+    }
+
+    public function testShouldConfigurePaginator(): void
+    {
+        $configurePaginator = fn (PaginatorConfig $config) => $config->setItemCountInPage(100);
+
+        $this->gridConfig->setConfigurePaginator($configurePaginator);
+
+        $this->assertNull($this->gridConfig->getPaginatorConfig());
+        $this->assertSame($configurePaginator, $this->gridConfig->getConfigurePaginator());
     }
 }