+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 - Filter fields in case-insensitive mode
 - [**BC**] Fix nullable types
 - [**BC**] Require mandatory parameters in constructor
+- Allow configuring paginator from the grid config
 
 ## 4.0.1 - 2022-03-25
 - just tag for a correct commit